requirements.txt
pytest
pytest-mock
dnspython
ldap3
mysql-connector-python
python-dotenv
rich